refactor: split lib and cli
